Added Ammo UI & Updated Ammo Equation
Added in the UI to track the player's ammo that's left and the current ammo they have in their gun. Also redid how the reload mechanic adds and subtracts bullets so that there's not a time in which there are more than the max amount of bullets available to the player without reloading.