Added Title and Game Over Screens
Added in the title and game over screens and their functionality. Game can be started and quit from the title screen, and restarted or quit from death. Only thing missing is the ability to look at the controls on the title screen. Fixed a bug where other objects could be taken if door used a Message Object as an unlocker.